// src/api/generated/models/CreateTagRequestDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateTagRequestDto = {
  /**
   * Title of the tag
   */
  name: string;
  /**
   * Color code of the tag
   */
  color: string;
  tasks?: Array<string>;
};